# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * متجري — SaaS perfume & cosmetics storefront platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MerchantRegister {
  storeName: string;
  /** Unique store URL slug */
  slug: string;
  email: string;
  password: string;
}

export interface MerchantLogin {
  email: string;
  password: string;
}

export interface Merchant {
  id: number;
  slug: string;
  storeName: string;
  email: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  /** @nullable */
  description?: string | null;
  accentColor: string;
  /** @nullable */
  bankTransferInfo?: string | null;
  /** @nullable */
  phone?: string | null;
  createdAt: string;
  /** @nullable */
  productCount?: number | null;
  /** @nullable */
  orderCount?: number | null;
}

export interface AuthResponse {
  merchant: Merchant;
  token: string;
}

export interface MerchantUpdate {
  storeName?: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  /** @nullable */
  description?: string | null;
  accentColor?: string;
  /** @nullable */
  bankTransferInfo?: string | null;
  /** @nullable */
  phone?: string | null;
}

export interface TopProduct {
  productId: number;
  productName: string;
  totalSold: number;
  totalRevenue: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface Order {
  id: number;
  merchantId: number;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  paymentMethod: string;
  isGift: boolean;
  /** @nullable */
  giftMessage?: string | null;
  /** @nullable */
  discountCode?: string | null;
  subtotal: number;
  total: number;
  status: string;
  createdAt: string;
}

export interface DashboardStats {
  ordersThisMonth: number;
  revenueThisMonth: number;
  totalOrders: number;
  totalRevenue: number;
  topProducts: TopProduct[];
  ordersByStatus: StatusCount[];
  recentOrders: Order[];
}

export interface ProductVariant {
  id: number;
  variantLabel: string;
  price: number;
  stock: number;
}

export interface ProductVariantInput {
  variantLabel: string;
  price: number;
  stock?: number;
}

export interface Product {
  id: number;
  merchantId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  imageUrls?: string[];
  /** @nullable */
  noteTop?: string | null;
  /** @nullable */
  noteHeart?: string | null;
  /** @nullable */
  noteBase?: string | null;
  /** @nullable */
  skinType?: string | null;
  /** @nullable */
  ingredients?: string | null;
  /** @nullable */
  batchExpiry?: string | null;
  isActive: boolean;
  createdAt: string;
  variants: ProductVariant[];
}

export interface ProductInput {
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  imageUrls?: string[];
  /** @nullable */
  noteTop?: string | null;
  /** @nullable */
  noteHeart?: string | null;
  /** @nullable */
  noteBase?: string | null;
  /** @nullable */
  skinType?: string | null;
  /** @nullable */
  ingredients?: string | null;
  /** @nullable */
  batchExpiry?: string | null;
  isActive?: boolean;
  variants: ProductVariantInput[];
}

export interface ProductUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  category?: string;
  imageUrls?: string[];
  /** @nullable */
  noteTop?: string | null;
  /** @nullable */
  noteHeart?: string | null;
  /** @nullable */
  noteBase?: string | null;
  /** @nullable */
  skinType?: string | null;
  /** @nullable */
  ingredients?: string | null;
  /** @nullable */
  batchExpiry?: string | null;
  isActive?: boolean;
  variants?: ProductVariantInput[];
}

export interface OrderItem {
  id: number;
  /** @nullable */
  variantId?: number | null;
  productName: string;
  variantLabel: string;
  quantity: number;
  priceAtOrder: number;
}

export type OrderDetail = Order & {
  items: OrderItem[];
};

export interface OrderListResponse {
  orders: Order[];
  total: number;
  page: number;
  pageSize: number;
}

export interface OrderStatusUpdate {
  status: string;
}

export interface OrderItemInput {
  variantId: number;
  quantity: number;
}

export interface OrderInput {
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  paymentMethod: string;
  isGift?: boolean;
  /** @nullable */
  giftMessage?: string | null;
  /** @nullable */
  discountCode?: string | null;
  items: OrderItemInput[];
}

export interface OrderConfirmation {
  orderId: number;
  total: number;
  storeName: string;
  storeSlug?: string;
}

export interface DiscountCode {
  id: number;
  merchantId: number;
  code: string;
  percentOff: number;
  isActive: boolean;
  createdAt: string;
}

export interface DiscountInput {
  code: string;
  percentOff: number;
  isActive?: boolean;
}

export interface DiscountValidation {
  code: string;
}

export interface DiscountValidationResult {
  valid: boolean;
  percentOff: number;
  code: string;
}

export interface StorePublic {
  slug: string;
  storeName: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  /** @nullable */
  description?: string | null;
  accentColor: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  bankTransferInfo?: string | null;
  createdAt: string;
}

export type ListProductsParams = {
category?: string;
q?: string;
};

export type ListOrdersParams = {
status?: string;
page?: number;
};