# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: متجري — SaaS perfume & cosmetics storefront platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerMerchant
      tags: [auth]
      summary: Register a new merchant account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchantRegister"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginMerchant
      tags: [auth]
      summary: Login merchant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchantLogin"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated merchant
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Dashboard Stats ─────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ─── Auth ───────────────────────────────────────────────
    MerchantRegister:
      type: object
      required: [storeName, slug, email, password]
      properties:
        storeName:
          type: string
        slug:
          type: string
          description: Unique store URL slug
        email:
          type: string
        password:
          type: string
    MerchantLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [merchant, token]
      properties:
        merchant:
          $ref: "#/components/schemas/Merchant"
        token:
          type: string
    Merchant:
      type: object
      required: [id, slug, storeName, email, accentColor, createdAt]
      properties:
        id:
          type: number
        slug:
          type: string
        storeName:
          type: string
        email:
          type: string
        logoUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        accentColor:
          type: string
        bankTransferInfo:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
        productCount:
          type: ["number", "null"]
        orderCount:
          type: ["number", "null"]
...