# lib/api-zod/src/generated/types/merchant.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * متجري — SaaS perfume & cosmetics storefront platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Merchant {
  id: number;
  slug: string;
  storeName: string;
  email: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  /** @nullable */
  description?: string | null;
  accentColor: string;
  /** @nullable */
  bankTransferInfo?: string | null;
  /** @nullable */
  phone?: string | null;
  createdAt: string;
  /** @nullable */
  productCount?: number | null;
  /** @nullable */
  orderCount?: number | null;
}